Moved examples to testing and removed old model generator.
