+++ glotaran/builtin/io/folder/test/test_folder_plugin.py
@@ -6,9 +6,9 @@ from typing import Literal
 import pytest
 
 from glotaran.analysis.optimize import optimize
-from glotaran.examples.sequential_spectral_decay import SCHEME
 from glotaran.io import save_result
 from glotaran.project.result import Result
+from glotaran.testing.sequential_spectral_decay import SCHEME
 
 
 @pytest.fixture(scope="session")
