+++ glotaran/deprecation/modules/test/test_project_result.py
@@ -5,8 +5,8 @@ import pytest
 
 from glotaran.analysis.optimize import optimize
 from glotaran.deprecation.modules.test import deprecation_warning_on_call_test_helper
-from glotaran.examples.sequential_spectral_decay import SCHEME
 from glotaran.project.result import Result
+from glotaran.testing.sequential_spectral_decay import SCHEME
 
 
 @pytest.fixture(scope="session")
